+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Allow customizing vault's etcd disk size.
+
 ## [14.15.0] - 2023-05-11
 
 ### Changed